nodeserver/worker/requirements.txt
absl-py==0.8.1
astor==0.8.0
certifi==2019.9.11
chardet==3.0.4
diffprivlib==0.2.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.3
h5py==2.10.0
idna==2.8
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
numpy==1.17.3
opt-einsum==3.1.0
Pebble==4.4.0
protobuf==3.10.0
psutil==5.6.3
requests==2.22.0
six==1.12.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-privacy==0.2.1
termcolor==1.1.0
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
dill==0.3.1.1
xgboost==0.90
sklearn==0.0
optuna==1.2.0
dill==0.3.1.1
scikit-learn==0.22.2.post1


